ビルドエラー対応 dbz.model（賦課関連のmodelをdbbに移動） dbz.business

+++ DB/DBZ/standard/model/src/test/java/jp/co/ndensan/reams/db/dbz/model/BemmeishaJohoModelTest.java
@@ -4,7 +4,7 @@
  */
 package jp.co.ndensan.reams.db.dbz.model;
 
-import jp.co.ndensan.reams.db.dbz.entity.basic.helper.DbT7003BemmeishaJohoEntityGenerator;
+import jp.co.ndensan.reams.db.dbz.entity.db.basic.helper.DbT7003BemmeishaJohoEntityGenerator;
 import jp.co.ndensan.reams.db.dbz.testhelper.DbzTestBase;
 import jp.co.ndensan.reams.uz.uza.util.db.EntityDataState;
 import static org.hamcrest.CoreMatchers.is;
@@ -63,8 +63,8 @@ public class BemmeishaJohoModelTest extends DbzTestBase {
 
         @Test
         public void 戻り値の弁明者枝番は_設定した値と同じ弁明者枝番を返す() {
-            sut.set弁明者枝番(DbT7003BemmeishaJohoEntityGenerator.DEFAULT_弁明者枝番);
-            assertThat(sut.get弁明者枝番(), is(DbT7003BemmeishaJohoEntityGenerator.DEFAULT_弁明者枝番));
+            sut.set弁明者枝番(DbT7003BemmeishaJohoEntityGenerator.DEFAULT_弁明者枝番.intValue());
+            assertThat(sut.get弁明者枝番(), is(DbT7003BemmeishaJohoEntityGenerator.DEFAULT_弁明者枝番.intValue()));
         }
 
         @Test
